tests/unit/test_charm.py
# Copyright 2023 Ghislain Bourgeois
# See LICENSE file for licensing details.
#
# Learn more about testing at: https://juju.is/docs/sdk/testing

import unittest
from unittest.mock import Mock, patch

import ops.testing
from charm import NginxOperatorCharm
from ops.model import ActiveStatus
from ops.testing import Harness


class TestCharm(unittest.TestCase):
    @patch("charm.KubernetesServicePatch", new=Mock)
    def setUp(self):
        ops.testing.SIMULATE_CAN_CONNECT = True
        self.addCleanup(setattr, ops.testing, "SIMULATE_CAN_CONNECT", False)

        self.harness = Harness(NginxOperatorCharm)
        self.addCleanup(self.harness.cleanup)
        self.harness.begin()

    def test_nginx_pebble_ready(self):
        # Expected plan after Pebble ready with default config
        expected_plan = {
            "services": {
                "nginx": {
                    "override": "replace",
                    "summary": "nginx",
                    "command": "nginx -g 'daemon off;'",
                    "startup": "enabled",
                }
            },
        }
        self.harness.container_pebble_ready("nginx")
        updated_plan = self.harness.get_container_pebble_plan("nginx").to_dict()
        self.assertEqual(expected_plan, updated_plan)
        service = self.harness.model.unit.get_container("nginx").get_service("nginx")
        self.assertTrue(service.is_running())
        self.assertEqual(self.harness.model.unit.status, ActiveStatus())
